Edited Emp and EmpClient

+++ TrainingSapient/src/com/sapient/OO/encapsulation/service/Emp.java
@@ -2,7 +2,7 @@ package com.sapient.OO.encapsulation.service;
 
 public class Emp {
 	private String name;
-	private int age;
+	private byte age;
 	private String busUnit = "RazorFish";
 	
 	public String getName() {
@@ -20,7 +20,7 @@ public class Emp {
 	public void setAge(int age) throws Exception {
 		if(age < 18 || age >= 60)
 			throw new Exception("Age must be between 18 and 60");
-		this.age = age;
+		this.age = (byte)age;
 	}
 	public String getBusUnit() {
 		return busUnit;
